docs: move quickstart to dedicated folder
