+++ quickstart/README.md
@@ -1,0 +1,3 @@
+# Connectors quickstart
+
+Scripts in this folder provide example on how to use connectors.
