+++ quickstart/google_sheets_2/README.md
@@ -1,4 +1,4 @@
-# OAuth2 Quickstart
+# Google Sheets (with OAuth2) Quickstart
 
 this small scripts is here to demonstrate the oauth2 workflow, applied to google.
 
@@ -12,7 +12,7 @@ go here: https://console.developers.google.com/
 
 Creez un nouveau projet. Dans la bibliotheque des API, activez 'Google Sheet'
 
-Allez dans la partie 'Identifiants'. Creez un nouvel identifiant OAuth. 
+Allez dans la partie 'Identifiants'. Creez un nouvel identifiant OAuth.
 En type d'applications, mettez 'Application Web'.
 En url de redirection, mettez 'http://localhost:34097', puis cliquez sur 'Creer'
