+++ sonar-project.properties
@@ -7,7 +7,7 @@ sonar.projectVersion=0.41.4
 
 # Path is relative to the sonar-project.properties file. Replace "\" by "/" on Windows.
 sonar.sources=./toucan_connectors
-sonar.coverage.exclusions=./toucan_connectors/install_scripts/**/*,./toucan_connectors/oauth2_connector/quickstart/**/*
+sonar.coverage.exclusions=./toucan_connectors/install_scripts/**/*,./quickstart/**/*
 sonar.test.inclusions=./tests
 
 # Encoding of the source code. Default is default system encoding
